Moved useCanvasContextOverride to component to prevent component reuse.
