+++ src/app/(pages)/pool/[pool-id]/participants/payout-scan/page.tsx
@@ -2,41 +2,12 @@
 
 import * as React from 'react'
 import { useState, useEffect, useRef, useCallback } from 'react'
-
-import BackCircleButton from '@/components/back-circle-button'
-import { useParams } from 'next/navigation'
+import { useParams, useRouter } from 'next/navigation'
 import { QrCodeCheckInData } from '@/types/qr'
 import { usePoolCreationStore } from '@/app/_client/stores/pool-creation-store'
-import { useRouter } from 'next/navigation'
 import PoolQrScanner from '../../_components/qr-scanner'
 import PageWrapper from '@/components/page-wrapper'
 import ScannerPageLayout from '@/components/scanner-page-layout'
-
-const useCanvasContextOverride = () => {
-    useEffect(() => {
-        const originalGetContext = HTMLCanvasElement.prototype.getContext
-
-        const customGetContext = function (
-            this: HTMLCanvasElement,
-            contextId: string,
-            options?: any,
-        ): RenderingContext | null {
-            if (contextId === '2d') {
-                options = options || {}
-                options.willReadFrequently = true
-            }
-            return originalGetContext.call(this, contextId, options)
-        }
-
-        // @ts-expect-error ts(2322) - This is a temporary fix to enable willReadFrequently for 2d context
-        HTMLCanvasElement.prototype.getContext = customGetContext
-
-        // Cleanup when unmounting the component
-        return () => {
-            HTMLCanvasElement.prototype.getContext = originalGetContext
-        }
-    }, [])
-}
 
 // Participant Check-in Preview
 export default function CheckInPage() {
@@ -48,7 +19,6 @@ export default function CheckInPage() {
     const params = useParams()
     const timerRef = useRef<NodeJS.Timeout | null>(null)
     const isProcessing = useRef(false)
-    useCanvasContextOverride()
 
     const router = useRouter()
